requirements.txt
Django==1.6
argparse==1.2.1
django-oauth2-provider==0.2.6.1
shortuuid==0.4.2
wsgiref==0.1.2
